[Feature-1119][admin,web,process] Add process console info
